add normal vector estimation
